fix: Repaint screen when view is too small
This catches many UI bugs where the content breaks across lines. But it's hacky to call such a dramatic function every loop, but I'm not sure a way around it. It appears to be broken behavior in Bubbles: Take the example Help bubble code, run it, shrink the terminal width, and you'll see the bugs.