fix: Don't stomp on quickly entered commands on app startup
If a command was executed before the Name of machine is resolved, the spinner would stop (and the selected command reset) when the message containing the resolved name would come in. Now, a specific name status message is sent when the name is determined. The end result is much more natural response in the UI =]